+++ lib/src/isolate_pool_executor_single.dart
@@ -226,6 +226,9 @@ class _IsolatePoolSingleExecutor implements IsolatePoolExecutor {
 
     return executor;
   }
+
+  @override
+  bool get isShutdown => _shutdown;
 }
 
 void _workerSingle(List args) {